// backend/src/obd/types/scan-event-type.enum.ts
export enum ScanEventType {
  ADAPTER_CONNECTED = 'ADAPTER_CONNECTED',
  ADAPTER_DISCONNECTED = 'ADAPTER_DISCONNECTED',
  VIN_READ = 'VIN_READ',
  VIN_READ_FAILED = 'VIN_READ_FAILED',
  DTC_READ = 'DTC_READ',
  DTC_READ_FAILED = 'DTC_READ_FAILED',
  VEHICLE_DATA_READ = 'VEHICLE_DATA_READ',
  DTC_CLEARED = 'DTC_CLEARED',
  DTC_CLEAR_FAILED = 'DTC_CLEAR_FAILED',
  CONTROL_UNIT_DISCOVERY_READ = 'CONTROL_UNIT_DISCOVERY_READ',
  ERROR = 'ERROR',
}